requirements.txt
yt-dlp>=2024.12.0
faster-whisper>=1.0.0
openai>=1.0.0
flask>=3.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
